# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BacklogMCP
  description: "Backlog SaaS\u3092Model Context Protocol (MCP)\u7D4C\u7531\u3067\u64CD\
    \u4F5C\u3059\u308B\u305F\u3081\u306EAPI"
  version: 0.1.0
paths:
  /api/projects/:
    get:
      tags:
      - projects
      summary: Get Projects
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u4E00\u89A7\u3092\u53D6\u5F97\
        \u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    project_service:\
        \ \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u7BA1\u7406\u30B5\u30FC\u30D3\u30B9\
        \uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u30D7\u30ED\u30B8\
        \u30A7\u30AF\u30C8\u4E00\u89A7"
      operationId: get_projects
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Get Projects
        '404':
          description: Not found
  /api/projects/{project_key}:
    get:
      tags:
      - projects
      summary: Get Project
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u3092\u53D6\u5F97\u3059\u308B\
        \u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    project_key: \u30D7\
        \u30ED\u30B8\u30A7\u30AF\u30C8\u30AD\u30FC\n    project_service: \u30D7\u30ED\
        \u30B8\u30A7\u30AF\u30C8\u7BA1\u7406\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\
        \u6027\u6CE8\u5165\uFF09\n\nReturns:\n    \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\
        \u60C5\u5831"
      operationId: get_project
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Project
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_key}/statuses:
    get:
      tags:
      - projects
      summary: Get Project Statuses
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u306E\u30B9\u30C6\u30FC\u30BF\
        \u30B9\u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\
        \u30F3\u30C8\n\nArgs:\n    project_key: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\
        \u30AD\u30FC\n    project_service: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u7BA1\
        \u7406\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\
        \nReturns:\n    \u30B9\u30C6\u30FC\u30BF\u30B9\u4E00\u89A7"
      operationId: get_project_statuses
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Project Statuses
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_key}/issue-types:
    get:
      tags:
      - projects
      summary: Get Project Issue Types
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u306E\u8AB2\u984C\u7A2E\u5225\
        \u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\
        \u30C8\n\nArgs:\n    project_key: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u30AD\
        \u30FC\n    project_service: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u7BA1\u7406\
        \u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\nReturns:\n\
        \    \u8AB2\u984C\u7A2E\u5225\u4E00\u89A7"
      operationId: get_project_issue_types
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Project Issue Types
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_key}/categories:
    get:
      tags:
      - projects
      summary: Get Project Categories
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u306E\u30AB\u30C6\u30B4\u30EA\
        \u30FC\u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\u30A8\u30F3\u30C9\u30DD\u30A4\
        \u30F3\u30C8\n\nArgs:\n    project_key: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\
        \u30AD\u30FC\n    project_service: \u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u7BA1\
        \u7406\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\
        \nReturns:\n    \u30AB\u30C6\u30B4\u30EA\u30FC\u4E00\u89A7"
      operationId: get_project_categories
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Project Categories
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_key}/milestones:
    get:
      tags:
      - projects
      summary: Get Project Milestones
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u306E\u30DE\u30A4\u30EB\u30B9\
        \u30C8\u30FC\u30F3\u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\u30A8\u30F3\u30C9\
        \u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    project_key: \u30D7\u30ED\u30B8\u30A7\
        \u30AF\u30C8\u30AD\u30FC\n    project_service: \u30D7\u30ED\u30B8\u30A7\u30AF\
        \u30C8\u7BA1\u7406\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\
        \uFF09\n\nReturns:\n    \u30DE\u30A4\u30EB\u30B9\u30C8\u30FC\u30F3\u4E00\u89A7"
      operationId: get_project_milestones
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Project Milestones
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_key}/versions:
    get:
      tags:
      - projects
      summary: Get Project Versions
      description: "\u30D7\u30ED\u30B8\u30A7\u30AF\u30C8\u306E\u767A\u751F\u30D0\u30FC\
        \u30B8\u30E7\u30F3\u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\u30A8\u30F3\u30C9\
        \u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    project_key: \u30D7\u30ED\u30B8\u30A7\
        \u30AF\u30C8\u30AD\u30FC\n    project_service: \u30D7\u30ED\u30B8\u30A7\u30AF\
        \u30C8\u7BA1\u7406\u30B5\u30FC\u30D3\u30B9\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\
        \uFF09\n\nReturns:\n    \u767A\u751F\u30D0\u30FC\u30B8\u30E7\u30F3\u4E00\u89A7"
      operationId: get_project_versions
      parameters:
      - name: project_key
        in: path
        required: true
        schema:
          type: string
          title: Project Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Project Versions
        '404':
          description: Not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/users/:
    get:
      tags:
      - users
      summary: Get Users
      description: "\u30E6\u30FC\u30B6\u30FC\u4E00\u89A7\u3092\u53D6\u5F97\u3059\u308B\
        \u30A8\u30F3\u30C9\u30DD\u30A4\u30F3\u30C8\n\nArgs:\n    backlog_client: Backlog\u30AF\
        \u30E9\u30A4\u30A2\u30F3\u30C8\uFF08\u4F9D\u5B58\u6027\u6CE8\u5165\uFF09\n\
        \nReturns:\n    \u30E6\u30FC\u30B6\u30FC\u4E00\u89A7"
      operationId: get_users
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response Get Users
        '404':
          description: Not found
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError